fix stab at getting header changes in. This is incomplete TBD:
       1) fix install so it installs the correct headers
       2) fix all files to use the new way to include files.

       This will however (for ease right now) just add $(QTDIR)/include/QT for each library
listed in the QT variable. I have copy and pasted the syncqt into syncqt.bat for now as well
but I will test on windows next. For the most part the only thing that (should) break with
this is anything that included a file that isn't in a library you're linking to will complain,
I've fixed everything in QTDIR/src that had this problem, notably it was qvariant.h and qevent.h.

[git-p4: depot-paths = "//depot/qt/main/": change = 136403]