decouple formats and document.
This simplifies the implementation greatly and will allow me to
reduce the complexity of the code quite a bit.

The only slight disatvantage might be that one cannot get the
format object from the format anymore. On the other hand these
do belong more into the document anyway from a conceptual POV,
so even this is probably cleaner now.

Next step will be to make the format objects children of the document.

[git-p4: depot-paths = "//depot/qt/main/": change = 138113]