remove the ref count from the Senders struct... use the active and orphaned flags (the same way as Connections)
change setActive(), resetActive(), setCurrentSender() and
resetCurrentSender() to save and restore the active flag.  inactive and
orphaned structures are automatically destroyed.

change resetActive() and resetCurrentSender() to compact the
Connections and Senders structs if the struct is inactive and a copy of
the vector is made... this helps cut down on wasted memory and keeps
the structures in a single memory block.

[git-p4: depot-paths = "//depot/qt/main/": change = 139785]