Renamed some "flag enum" types to be more consistent (i.e. the enum type associated to a flag type through Q_DECLARE_FLAGS).
Usually, the enum type isn't used directly; only the flags type
is. Still, sometimes we do use it, and then it's important to have
a good name. Also, the documentation shows that name.

The rule I've tried to follow were:

	1. If the flags type is a plural (e.g. Options, Actions, Flags),
	   make the enum type a singular (e.g. Option, Action, Flag).

	2. Otherwise, make the enum type by appending "Flag" to the
	   flags type (e.g. Alignment is the flags type, AlignmentFlag
	   is the enum type).

Notice that "Flag" is a singular. This is because an enum of that
type can only store one flag at a time (whereas a variable of the
"flags" type can store many).

[git-p4: depot-paths = "//depot/qt/main/": change = 141390]