- make all actions in QToolBar index based - add support for widgets in the tool bar (also action based).
the gory details involves storing a list of items that tells about the
action/widget in the correct order.  we also do our best to keep the
QWidget::actions() list in the same order as the toolbar list.

this API should be enough to allow the ide people to write a nice
action editor for QToolBar.

[git-p4: depot-paths = "//depot/qt/main/": change = 142155]