make shortcuts in popup menus work again (the compat functions).
The trick is, that an action needs a widget ancestor to install an event filter on 
in order to receive shortcut events. If you create an action without a parent (or you 
create a menu without a parent), we cannot have shortcuts. For that reason it is 
better to not offer convenience functions like addAction(...) in QMenu that also take
a shortcut. If you want a shortcut, you can call setShortcut() on the action, or create an 
action yourself.

QAction now generates a warning message if you set a shortcut and there
is no widget parent.

[git-p4: depot-paths = "//depot/qt/main/": change = 142358]