Fix some painting errors on windows.
It appears that the HDC has a black pen set when we return from
GDI+. We therefore need to trigger update on the pen before drawing a
rectangle. GDI will otherwise draw rectangles one pixel too large.

We introduce the updateInternal recursive call in QPaintEngine to
maintain integrity for the case where one update flags another
state as dirty.

The above error in the pen caused me to make the wrong assumption
about the pen vs advancedMode, which is now fixed.

[git-p4: depot-paths = "//depot/qt/main/": change = 144167]