when using XLFD fonts, we need to make sure that the size we select actually supports the script that we want.  XLFD fonts are really strange in that each size has an associated encoding, and you can have multiple encodings per size.
this causes a problem when you try to load a font for a script in a
size that doesn't have the proper encoding.  this is quite common: the
misc-fixed font normally comes in 12 pixels for latin, but normally
only 14 or 16 pixels for asian.

this fix makes sure that we find a size that supports the script
(encoding) we want, giving larger penalties to fonts that have a larger
size than the one requested.  this allows us to load the closets
possible size.

reported by Trond during 3.3.2 package testing

Reviewed (twice!) by Lars.

[git-p4: depot-paths = "//depot/qt/main/": change = 144197]