introduce QThreadData, an internal class which holds thread-specific data (event loop, post event list, tls). threads created by QThread (and the thread that creates QApplication) will have an associated QThreadData, threads created using other toolkits or platform APIs will NOT get a QThreadData.
this is a port of change 144528 from 3 to main:

provide a safety mechanism when trying to use QThreadStorage from
non-QThread threads.  instead of using a global tls for all such
threads, spit out a warning and do nothing (much better than data races
and crashes)

QThreadStorage is used in QRegExp, so we need to be able to handle the
case where QThreadStorage doesn't work (e.g. in non-QThread threads).
essentially, the regexp engine cache is unused in such threads.

fixes problems reported by FirstLogic, task 49393

[git-p4: depot-paths = "//depot/qt/main/": change = 145909]