Streamline the rendering pipeline.
In QPaintEngine::updateInternal we detect when features must
be emulated by QPainter. For each and every draw call we
then check if emulation is required and call ONE function which
does it all (QPainter::draw_helper). To avoid overhead, TxTranslate
is handled in each draw call as a special case.

Combined with path outlining we now get penwidth xform and
alphapens on all platforms, plus it opens for antialiased lines
on X11.

This change is bound to break some painting code, and perhaps
compilation, but I will look into those on monday.

[git-p4: depot-paths = "//depot/qt/main/": change = 151720]