- give the QConnectionList members better names
- optimize QConnectionList by allowing it to reuse unused slots in the
connections list... this also involves an invariant flag that disallows
reuse. this is needed in QMetaObject::activate().  if the next
connection to be activated is disconnected, and another connection is
added which identical to the one that was just connected, we do *NOT*
want this new connection to be activated.  autotest already added

- optimize QMetaObject::activate(): do not use QHash::values() (which
returns a QList).  we want to avoid creating the list, so we use a
recursive function (!!) instead. hats off to jasmin for the
idea... this works extremely well, and improves performance quite a
bit.

[git-p4: depot-paths = "//depot/qt/main/": change = 156808]