Suppose we press the mouse middle button: 	void QToolBarHandle::mousePressEvent(QMouseEvent *event) 	{ 	    if (event->button() != Qt::LeftButton) 	        return; 	    Q_ASSERT(parentWidget()); 	    Q_ASSERT(!state); 	    state = new DragState; Obviously "state" will not be set.
Then move the mouse without releasing the button:
	void QToolBarHandle::mouseMoveEvent(QMouseEvent *event)
	{
	    Q_ASSERT(state != 0);
Something's wrong here. This function expects "state" to be set.

I think the latter function shouldn't expect "state" to be set.
Is this the correct fix?

By the way, there are no visual cues when dragging the toolbar.
This is really not user-friendly, especially since the toolbar
does not move in an intuitive way.

[git-p4: depot-paths = "//depot/qt/main/": change = 160139]