Delete zlib 1.2.1.
This is insane! From the zlib-announce mailing list:
	zlib users,
	Though it has not been updated on http://www.zlib.org/ yet,
	the new version of zlib, version 1.2.2, is available here:
		http://www.zlib.net/
	All users of zlib 1.2.1 should update to 1.2.2. zlib 1.2.2
	eliminates a potential security vulnerability and fixes a
	bug in decompression.
This message had been sent on October 16 but http://www.zlib.org/ 
hasn't been updated yet!

Anyway, this isn't really an issue in our case since zlib 1.2.2
fixes a denial of service vulnerability in zlib 1.2.1:
http://cve.mitre.org/cgi-bin/cvename.cgi?name=CAN-2004-0797
which we had already patched in change 151654.

[git-p4: depot-paths = "//depot/qt/main/": change = 161736]