manual integration: Integrate 161743-161745 from 3 to main:
Delete zlib 1.1.4.

This really is an old unmaintained version. Security issues that
are fixed in the 1.2 series are not fixed in the 1.1 series and
are getting harder to fix. Upgrading to zlib 1.2.2 looks like
the best solution to me, even if this upgrade happens in a
patch release.

Add zlib 1.2.2.

Fixes #59702.

Trolltech modifications to zlib.

[git-p4: depot-paths = "//depot/qt/main/": change = 161746]