With regard to the no comment in change #163113
The order of the QAtomic and the value union was swapped to prevent a crash on some version of VC6 on Windows 2000. This is mostly likely a compiler bug that we suspect is related to a 8 byte union being incorrectly aligned on a 4 byte boundary. The position of the pointer inside the union probably does not make a difference but it worked and so we left it that way.

[git-p4: depot-paths = "//depot/qt/main/": change = 163161]