QTextStream::readLine() is documented to _not_ include the end-of-line terminator, but it did nevertheless. This fix makes sure it does not include the terminator.
The fix is not perfect; if a file ends with a '\r', the last
line will include this character. Also, reading a Mac OS 8/9
text file with 'foo\rbar' still returns this as one line 
(including the CR). The plan is to rewrite QTextStream's
internals in Qt 4.1, so that we can fix this efficiently.

[git-p4: depot-paths = "//depot/qt/main/": change = 163199]