Renaming: socketState() and processState() become state() again. socketError() and processError() become error() again.
Background: As part of the QIODevice changes we've been through in main,
we decided to find longer, better names to describe the state and error
values of different devices. socketState() was preferred instead of
state(), as Qt already defines too many state types. This especially
became a problem when subclassing from multiple classes, where .state()
does not say enough about the context in which it was called, rendering
the code hard to read.

An additional problem was that the fetcher methods returned ints in Qt 3. The state enum
was called "State" in all classes; confusing both when
reading documentation and when using the feature (see QIODevice::State as opposed
to QSocket::State in Qt 3). The fix in Qt 4 was to rename the enums to
SocketState. This makes docs easier to read, and disambiguates states in
different classes. However, the renaming of state() to socketState() was
a mistake because the variable name itself brings all the context needed to
read the code. For the case of multiple inheritance, we fell into the trap
of awarding a few for the burden of many.

This change was agreed on after a discussion with Lars, Matthias, Jasmin,
Sam and I, before Beta 1 froze.

[git-p4: depot-paths = "//depot/qt/main/": change = 163721]