fix a few bugs in the post event mechanism:
- when QCoreApplication is destroyed, don't leak left over posted
events, they should be deleted (and the receiver's postedEvents count
decremented)

- when sending posted events for a specific receiver, we should use
mainData() instead of QThreadData::current() if the receiver has no
thread.  As a sanity check, make sure that sendPostedEvents() with a
receiver is called by the receiver's owner thread.

- make sure we update the receiver's postedEvents counter when removing
a specific event

fixes a bug reported on the qt4 feedback list

[git-p4: depot-paths = "//depot/qt/main/": change = 163825]