Make it possible to link plugins statically into an application. Due to linkers removing unreferenced symbols, this requires explicit referencing in the application. Sucks, but nothing we can fix.
Here's how to do it:
1. compile your plugin as static library with -config=staticlib
2. link the .a into your application
3. reference the plugin with Q_IMPORT_PLUGIN(name) somewhere
in your application.

"name" is the same name that is used for Q_EXPORT_PLUGIN, for 
example "WindowsStyle".

The mechanism makes it possible to use a static Qt and still have
e.g. accessibility. Also designer could be linked statically without
changing too much how it works with plugins.

[git-p4: depot-paths = "//depot/qt/main/": change = 164374]