Name changes:
1. viewportRectForIndex() -- used to be itemViewportRect()
2. sizeHintForIndex()     -- used to be itemSizeHint()
3. sizeHintForRow()       -- used to be rowSizeHint()
4. sizeHintForColumn()    -- used to be columnSizeHint()
5. rootIndex()            -- used to be root()
6. setRootIndex()         -- used to be setRoot()
7. entered()              -- used to be itemEntered()

Names 1 and 2 are necessary to avoid a clash with functions of the same
name in the convenience classes. Names 3 and 4 need to be changed for
consistency with 2. Name 5 and 6 don't need to be changed, but it seems
appropriate to do so (to prevent confusion or a future name clash). Name
7 is renamed to be consistent with other signals (e.g., pressed(),
clicked(), doubleClicked()).

[git-p4: depot-paths = "//depot/qt/main/": change = 164606]