implemented the new scalable QIcon, QIconEngine, and QIconEnginePlugin. The documentation won't win any Pulitzer prize, but at least there is documentation. Not sure what to do with Qt::IconSize. Maybe this should go entirely and be replaced by QSize. If the style allowed to query for preferred icon size for different roles on different systems, that might be good enough.
There is currently no cache in the default QPixmapIconEngine, and I
don't really know whether we need one or not. Probably we do (for slow
remote connections). QPixmapIconEngine should then simply use the
global QPixmapCache.

Another possible modification is in QStyle: replace drawPixmapItem with
drawPixmapIcon and call QIcon::paint() instead of first creating a pixmap.
This will be faster once we have a QSvgIconEngine :-)

[git-p4: depot-paths = "//depot/qt/main/": change = 165166]