Rename QBezier::offsetted() -> QBezier::shitted.
Introduce a curve segmentation threshold to make it possible
to increase the resolution of the stroked path based on
the scale factor of the device it will be rendered to. This
makes it possible to draw ellipses at (0, 0, 1, 1) which 
are completly round ;-)

Also start using QBezier::shifted in 
QPainterPathStroker::strokeCurve so that the precision of
strokes become better.

[git-p4: depot-paths = "//depot/qt/main/": change = 167236]