Use different names for member variables and constructor arguments that initialize them.
That said, it looks like there is a misunderstanding what
q*_p.h headers are for. They are for declaring types and
instances that are private to Qt but shared by many Qt
sources.
That's apparently not the case with Command and CommandType,
so these should be declared in qlineedit.cpp instead of
qlineedit_p.h.

[git-p4: depot-paths = "//depot/qt/main/": change = 168350]