check last window closed only if type == Qt::Window || ((type == Qt::Dialog || type == Qt::Tool) && !q->parentWidget())
Not sure whether this is the right thing to do, but it is in line with what Qt 3 did. Ideally we would only 
treat Qt::Window as windows, and have those who write special apps (stuff that only is a tool or a dialog) do
something different (like calling exit themselves). Hmmm... maybe this is exactly what we should do?

[git-p4: depot-paths = "//depot/qt/main/": change = 168515]