p4i integration Integrate 169778 from 4.0 to main: Fixed checkmark drawing in menus in motif style
There was still old logic there that allowed a menu to be checkable or not, 
with the idea that motif style wouldn't draw a check column on a non-checkable
menu. The property was changed in QStyle::polish(QWidget *w). This was
broken, and it didn't even work. Motif didn't show any check columns.

We removed QMenu::checkable and instead added
QStyleOptionMenuItem::menuHasCheckableItems. This fixes the usability
issue that if you have checkable actions in a menu and you develop on
Windows or Mac, you wouldn't even see those on Unix in Motif style
unless you called menu->setCheckable(true). Arglll.

Fixed documentation of QAction::checkable

Done with Marius I. 

[git-p4: depot-paths = "//depot/qt/main/": change = 169780]