p4i integration Integrate 169801 from 4.0 to main: hotfix: Fixed a possible crash in Resource/FormBuilder. The Resource/FormBuilder is the replacement we have in Designer 4 for the QWidgetFactory. For now I'm trying to be very *strict* in the Resource. I mean if the property type is unknown, or the property name is not stored in the QMetaObject the Resource will asserts.
So essentially the Resource needs a valid QMetaObject. Unfortunatly many widgets
like (SpacerWidget or QLayoutWidget) are internals. The Resource doesn't really
know about them, and actually it doesn't even link with these classes.
We have to fake the properties.

The trick I'm using here is to create a Fake<Widget> class in the Resource. This class
is a placeholder. And it is not supposed to be used. The only thing we want is to
store the informations about the properties and the enums in a *valid* QMetaObject instance.

reviewed by Jacek

[git-p4: depot-paths = "//depot/qt/main/": change = 169802]