p4i integration Integrate 170558 from 4.0 to main: Change 170558 by twschulz@twschulz-tick on 2005/03/04 15:13:50
Fixes:    Consistency on our exported C-functions
Task:     none
RevBy:    Jasmin
AutoTest: none
Details: We have a fair amount of C-functions that are usable by customers but they are not documented, named consistently, or properly exported. This is an attempt to clean this up. The functions have been renamed to match the qt_XXX_ style. Platform specific functions are prefixed qt_x11, qt_win, or qt_mac as appropriate and they are now documented. Jasmin to send out an email with the extra information shortly.

[git-p4: depot-paths = "//depot/qt/main/": change = 170564]