p4i integration Integrate 170666 from 4.0 to main: Fixes: Size of windows changes in workspace mode when switching modes or restoring Task: None RevBy:    Lars AutoTest: None Details:
Guys, it's really not that hard :-) pos() in toplevel mode includes the
frame, for a workspace window you must look at
parentWidget()->pos(). That's all. But you _cannot_ pass a workspace's
parent widget as widget and expect a decent geometry. While the
position is correct, the size is bigger (it _includes_ the frame).

[git-p4: depot-paths = "//depot/qt/main/": change = 170667]