p4i integration Integrate 171442 from 4.0 to main: Fixes:    remove the hack on X11 for color0 and color1 Task:     none RevBy:    trond AutoTest: none Details:
color0 and color1 were given special values (alpha == 0x01) to help
QColormap to return the correct pixel values... this broke the painter
emulation logic, so we need to remove it and solve it in a cleaner way.
now, we threshold all colors when drawing on to a 1bpp destination.  if
the color is close to white, we draw using color 0 (clear), otherwise
we draw with color 1 (set).

[git-p4: depot-paths = "//depot/qt/main/": change = 171443]