p4i integration Integrate 172121 from 4.0 to main: Change 172121 by eblomfel@eblomfel-SpareRibs on 2005/03/16 17:36:34
Fixes:    Fixed undo for subsequent inserts on a Q3TextEdit
Task:     30285
RevBy:    Lars
AutoTest: 112290
Details: 
insert() in Q3TextEdit assumed that all insertions in a sequence directly appended each other, and 
stored them all in a single command on the undo stack. I added a test so that new insertions are only
placed in the same command as the previous if it is actually inserted directly after the previous.

[git-p4: depot-paths = "//depot/qt/main/": change = 172122]