p4i integration Integrate 172196 from 4.0 to main: Fixes:    fix QMainWindow::iconSize and QToolBar::iconSize Task:     68101 RevBy:    trond AutoTest: 172183 172184 172191 Details:
the iconSize property in QMainWindow and QToolBar should be completely
orthonongal to each other.  there should also be a clear propagation
method as well: the toolbar should use the mainwindows's icon size if
the toolbar's icon size has not been explicitly set.

on style change, we need to make sure that we update the icon size in
both mainwindow and toolbar, but only if they have not been explicitly
set

[git-p4: depot-paths = "//depot/qt/main/": change = 172197]