p4i integration Integrate 173125 from 4.0 to main: Remove more duplicate header inclusions.
Header inclusion is a mess in Qt 4, much worse than Qt 3.
We really need some automatic tests to enforce style and
improve readability.

I've used similar scripts to find duplicates in Qt header
files:
	fgrep 'qt/4.0/src/*/*/*.cpp \
		| sed -e 's/"\(.*\)"/\1/; s/<\(.*\)>/\1/' \
		| sort \
		| uniq -d
Unfortunately there are wrong positives (from comments and
example code in documentation) and it needs more work anyway,
but that's the kind of script that could be run on a daily
basis on the Qt code base.

[git-p4: depot-paths = "//depot/qt/main/": change = 173126]