p4i integration Integrate 173638 from 4.0 to main: Fixes:    QObject::thread() Task:     68363 RevBy:    andreas AutoTest: 173571 173569 173567 173565 Details:
a thread value of zero no longer means "the main thread", it means "no
thread"... this means that Q(Core)Application creates a QThread for the
main() thread... this prepares us for the ability to move objects
between threads.

note: one side effect of this change is that posted events for objects
with no thread affinity will be processed by the main thread.

[git-p4: depot-paths = "//depot/qt/main/": change = 173639]