p4i integration Integrate 173791 from 4.0 to main: doc: Move all child widget variable declarations to .h file. Rational: A common mistake for beginners is to type
	QComboBox *foo;

in the header file and

	QComboBox *foo = new QComboBox(this);

in the class constructor. Then, the 'foo' data member is
uninitialized, and the program will crash if 'foo' is used
outside of the constructor. Therefore I think that it's slightly
better to consistently declare variables in the .h file. It also
looks cleaner in the .cpp file. --Jasmin

[git-p4: depot-paths = "//depot/qt/main/": change = 173792]