p4i integration Integrate 173829 from 4.0 to main: Fixes:    prepare for the ability to move objects between threads Task:     68363 RevBy:    harald AutoTest: existing qtimer test Details:
in order to be able to move objects between threads, we need to have
application wide timer ids... the best way to do this is to make the
QAbstractEventDispatcher interface take the timer id as a parameter.

also, we need a way to list the currently registered timers (so that we
can later register them with another dispatcher when an object moves to
another thread).

[git-p4: depot-paths = "//depot/qt/main/": change = 173831]