p4i integration Integrate 174256 from 4.0 to main: Fixes:    Colors of monochrome images. RevBy:    Trond & Lars Details:
By default a monochrome image has no colortable. When converted to a 32
bit it gets 0=black and 1=white. This is the standard that image apps
uses and also follows the pattern you would see if when converting a 32
bit to a 1 bit.

Fix also the case where ARGB32_Premultiplied is converted to mono as we
did not compensate for the premultiplied rgb values.

[git-p4: depot-paths = "//depot/qt/main/": change = 174257]