p4i integration Integrate 175059 from 4.0 to main: Fixes:    better rules to find locale RevBy:    could someone please review this? Lars? Details:  Use correct precendence rules.           LC_ALL has precedence over LC_NUMERIC.           LC_NUMERIC has precedence over LANG. http://www.opengroup.org/onlinepubs/007908799/xbd/envvar.html http://publib.boulder.ibm.com/infocenter/pseries/topic/com.ibm.aix.doc/aixprggd/nlsgdrf/locale_env.htm
Also LC_NUMERIC just characterizes numeric formats. It doesn't
characterize the whole current locale. As such, it's not clear
whether it should have precedence over LANG as far as these
functions are concerned:
	QLocale::country()
	QLocale::language()
	QLocale::name()
I think it shouldn't - in that case some of the QLocale source
code needs to be rewritten.

[git-p4: depot-paths = "//depot/qt/main/": change = 175060]