p4i integration Integrate 176237 from 4.0 to main: Change 176237 by eblomfel@eblomfel-SpareRibs on 2005/04/20 16:01:46
Fixes:    Update problem in Supporter
Task:     none
RevBy:    Trenton, Håvard
AutoTest: Supporter
Details:
Q3Listview calls setUpdatesEnabled() in lots of places. If it's inside of a 
stacked layout, when it's shown/hidden setUpdatesEnabled() is called on it 
by the layout. However Q3ListView sets updates to be forced disabled if 
Q3ListView implicitly had updates disabled. This caused problems when we 
implicitly enabled updates again. The trick is to simply set the 
WA_UpdatesDisabled attribute instead of the function call. 

[git-p4: depot-paths = "//depot/qt/main/": change = 176238]