p4i integration Integrate 177143 from 4.0 to main: Change 177143 by qt@qt-lusa on 2005/04/26 12:33:30
Username: eblomfel
Fixes:    Windows weren't created on Windows ME
Task:     none
RevBy:    Andreas
AutoTest: none
Details:
The char-pointer sent to RegisterClass was pointed to internal 
data in a QByteArray object returned by the toLatin1() call. 
This object seemed to be deleted immediately, and in debug 
mode the string was overwritten with debug info. Therefore 
the window classes were never registered in debug mode. 
The fix was to copy the QByteArray object out into a local 
variable.   

[git-p4: depot-paths = "//depot/qt/main/": change = 177145]