p4i integration Integrate 177400 from 4.0 to main: Fixes:    atomic on the arm Task:     65058 RevBy:    none, trust me (lars in spirit) AutoTest: existing tests Details:
in the interest of space, use a trick on the arm that is similar to the
one we do on parisc: use a global array of locks instead of a single
global lock for all atomic operations...

this is all done as function calls, which gives us the possibility of
providing a binary compatible, optimized drop-in replacement for ARMv6,
which implements load/store exclusive (similar to lwarx/stcwx. on ppc).

[git-p4: depot-paths = "//depot/qt/main/": change = 177401]