p4i integration Integrate 179833 from 4.0 to main: Change 179833 by eblomfel@eblomfel-SpareRibs on 2005/05/12 12:16:49
Fixes:    RCC generated invalid files when there were no resources
RevBy:    Trust me
Details:
Fix RCC to quit with an error code when it has no resources to output.
It should not generate the output file at all, so that it will be run
automatically on every new execution of 'make'. This is a trust me 
as far as the code goes, but the feature was discussed with Lars.

[git-p4: depot-paths = "//depot/qt/main/": change = 179834]