p4i integration Integrate 187912 from 4.0 to main: Change 187912 by eblomfel@eblomfel-SpareRibs on 2005/07/19 10:18:11
Fixes:    Sliders in classic windows style not drawn differently in disabled state
Task:     62663
RevBy:    Trenton
AutoTest: none
Details:  
There was no code for drawing disabled sliders in the Windows style. There is now.
Windows common controls draws disabled track bars identical to enabled 
track bars except that the inner polygon of the handle is drawn in a dithered
pattern. There's also a bug in the drawing on native Windows, but I decided
not to reproduce that :)

[git-p4: depot-paths = "//depot/qt/main/": change = 187914]