p4i integration Integrate 188152 from 4.0 to main: Change 188152 by eblomfel@eblomfel-SpareRibs on 2005/07/20 15:54:00
Fixes:    Drawing of default buttons in windows style
Task:     77334
RevBy:    Trust me
AutoTest: none
Details:  
Change #188054 was not the right fix. The default buttons were being
drawn too large, but only when sunken. Change #188054 was a fix for
the actual bug and also a work-around for a bug in the painting system.
The painting bug has been fixed, and hopefully this change will live 
longer than the previous one.

[git-p4: depot-paths = "//depot/qt/main/": change = 188153]