Fixes:    configure tests looking in the wrong locations for libs/headers Task:     yes AutoTest: no thanks Details:
on systems that support both 32 and 64 bit binaries, we often look in
the wrong directories when trying to find libraries (i.e. .../lib
instead of .../lib64)... this is complicated when forcing a particular
output on such platforms (i.e. linux-g++-32 to force 32bit output on
64bit machines)

the solution is to stop doing the configure tests in shell scripts and
instead do compile/link tests... this means we have to do the tests
after qmake is built.

this change adds the generic config.tests/x11/compile.test shell script
as a trampoline to build all of the X11 tests (which are also converted
in this change)

config.tests/unix/... is next

[git-p4: depot-paths = "//depot/qt/main/": change = 194002]