Fixes:    xformed span generation in raster engine. Task:     85128 and a few others... RevBy:    Done with Eskil AutoTest: covered by lance tests
Details: We had a few concepts a bit wrong here previously, but I think
we got it now...

Important rule is: Pixels are defined by the four mathmatical points
lying enclosing them.

Normally we snap pixels to the top left point and work with it. This is
wrong for transformation since the type of transformation affects which
way to snap. Scale (1, -1) for instance means that we snap to bottom
right point instead. 

When deciding on the value for any pixel we fill in, we do this by
looking at the center of the pixel (thus the +0.5) and map this as best
as possible to log coords.

In normal xform we snap to top left in texture coords (snap direction
is now in the matrix because of the +0.5). 

In smooth xform we look at four pixels and treat each of these pixels
as a rectangle spanning 1x1, with the pixel center on the point. This
means that the texture coordinate system is off by -0.5, -0.5 which we
need to compensate for.

[git-p4: depot-paths = "//depot/qt/main/": change = 194322]