Fixes:    Printing searchable files when printing to file on Windows && clean up dead code Task:     85823 Details: Two pieces of stuff:
1. When printing latin1 strings, we will now convert the glyph indices to text before sending the data to ExtTextOut. 
Using e.g. PDF printer drivers this will make the actual text be embedded in the resulting file. We are not doing the
same for non-latin1-characters as this can break in some cases. For non-latin1-chars the glyph indices are printed
directly. Note: This is only supported for HighResolution-printing. ScreenResolution-printing still goes via paths in the
paint engine.

2. Clean up some dead code in the text output function. Transformed text is now always converted to paths and 
passed through the paint engine, so we don't need all the transformation code in draw_text_item_win().

[git-p4: depot-paths = "//depot/qt/main/": change = 194675]