Fixes:    Small optimization to setPen/setBrush
Details: In live applications its often hard to know if the right
pen/brush is set or not so the tendency is to reset pen/brush
frequently. If this is done once for each drawXXX call updateState
takes around 18% of the draw call.  By adding a simple check we reduce
this to <1%.

It should also be said in my test case doing the checks is slightly
faster than doing the qAtomicAssign (0.6% vs 1%) which is performed by
QPen::operator=

[git-p4: depot-paths = "//depot/qt/main/": change = 195107]