Fixes:    crash on X11 when asking for QStyle::standardPalette() before creating QApplication RevBy:    twschulz AutoTest: tst_qapplication Details:
QX11Info is initialized as part of the QApplication constructor, so we
can only use it after QApplication has been constructed. In this
particular case, we use the win2000 colors on >8bpp displays or if QApp
hasn't been created yet.

Fix change 197273 a little bit to remove duplicate code, and to move
the palette setting code together in one place

[git-p4: depot-paths = "//depot/qt/main/": change = 197327]