p4i integration Integrate 199223 from 4.1 to main: Fixes:    crash when returning 0 from QWidget::paintEngine() RevBy:    Andreas
Details: There are quite a few people who are returning a 0 paint
engine to bypass Qts painting all together. This is so that they have
direct access to the HDC without any buffering or what-not, such as
media players. The change is a few lines but its pretty much only
safe guarding access to the paintEngine pointer that the widget
returns.

[git-p4: depot-paths = "//depot/qt/main/": change = 199224]