p4i integration Integrate 209233 from 4.1 to main: Fixes:    undesired image scaling in QText... RevBy:    Done with Simon
Details: If we get an image with a dpi on windows we preserve this so
that the image is scaled according to its specified dpi based
width/height while on X11 we ignore this alltogether. Of course the
image is mostly specified in 72dpi which on windows causes a scale factor
of 96/72 so windows gets a nasty scale while X11 looks good. 

Again resolution independence doesn't work... The code we changed
was there for printing, in which case only the difference between
screen/printer is interresting. This makes it impossible to
do proper typography on windows, but that never worked on X11 anyway
due to QPixmap::fromImage not compensating for resolution difference
in the image and the X server.

[git-p4: depot-paths = "//depot/qt/main/": change = 209236]