p4i integration Integrate 211291 from 4.1 to main: Fixes:    extremly Slow dotted lines Task:     102122 RevBy:    Trond Details:  Its still not optimal, but significantly better.
	Dotted lines are drawn by converting a line into a series of
moveTo/lineTo's. If the line is long, this means a ton of segments
and an awfully complex polygon to render in the end. Since the
segments, where both move and line -to are outside we can skip
the emission. This speeds it up dotted line drawing of long lines 
to an acceptable level...	

[git-p4: depot-paths = "//depot/qt/main/": change = 211292]