p4i integration Integrate 212777 from 4.1 to main: Fixes:    Application crashes when clicking mouse button while dragging Task:     103277 RevBy:    lars
Details: if we start a new drag before the previous finished (i.e. the
target hasn't gotten the data yet), we'd end up asserting... we fix
this by creating transactions for each XdndDrop we send. when the
target requests the data, we use either data from one of the pending
transactions or the current drag data (determined by the timestamp of
the SelectionRequest).

       note that there's a bit of extra code to support previous (broken) Qt
       versions, which didn't use the correct timestamp when requesting data
       as specified in the Xdnd spec.

[git-p4: depot-paths = "//depot/qt/main/": change = 212779]