Fixes:    (Partially) Qt building where srcdir or builddir contains spaces Task:     None AutoTest: __compile_qt__ Details:  There were LOTS and LOTS of places where the variables weren't properly quoted. So I quoted them all. And then I found some more. And then I realised I had to refactor QMAKE_VARS completely because it was simply completely unusable for the task (it could not handle quotes inside quotes).
So: ./configure now has all but two(*) variables properly
quoted so as to be able to run the configure script on a
shadow build containing spaces. The config tests have been
updated as well.

qmake/Makefile.unix had to be refactored to allow for quoted
filenames (actually, backslash-escaped).

This does not break the current Qt build for me, but it still
is not enough. The remaining two places for quoting the
variables break a normal build.

[git-p4: depot-paths = "//depot/qt/main/": change = 218152]