Fixes:    Adapt Q3SocketDevice to the new QIODevice Task:     None, found by Trenton and tst_q3urloperator AutoTest: Existing tst_q3urloperator Details:  QIODevice now defaults to having an internal buffer. That makes no sense in Q3SocketDevice, so set the Unbuffered flag. (The same change had to be made for QUdpSocket in change 220776).
Also, do not close the socket when read(2) returned 0, as that
will get the QSocketNotifers in Q3Socket trying to poll a
closed file descriptor. I should test if
connect-close-reconnect still works in Q3Socket...

Finally, call the base implementation in
Q3SocketDevice::bytesAvailable. We tell our customers to do
it, so we may as well take our own poison :-)

[git-p4: depot-paths = "//depot/qt/main/": change = 221539]