Fixes:    Add wildcard signal connection and signal id tracking Task:     None AutoTest: None Details:  These two features are required by QtDBus (and these simple changes shave off quite a bit of ugly code in QtDBus, not to mention 3% of the CPU spent in the signal spy hooks).
One feature is the saving of the signal IDs in QObjectPrivate
when the signal is emitted, along with the sender
(QObject::sender). There is no public way to access this.

The second feature is to allow QMetaObject::connect to take -1
as its signal meta-method ID, which means "any signal" when
delivering the connection to the slot.

[git-p4: depot-paths = "//depot/qt/main/": change = 224185]