p4i integration Integrate 229423 from 4.2 to main: Add error checking to QMacPasteBoard.
We were getting a lot of "Failure to create pasteboard" errors when running the
autotests on minime, usually followed by a crash. What is happening is probably
that the pasteboard does not work when the user is not logged in at the screen.

In any case Qt should not crash if the pasteboard creation fails, this submit fixes
that.

[git-p4: depot-paths = "//depot/qt/main/": change = 229424]