p4i integration Integrate 232095 from 4.2 to main: Fix off-by-one painting errors in the painting engine. Trying out a new way of accounting for the inverted Quartz coordinate system.
Basically we add a little "delta" value to each y coordinate when stroking, and keep
the coordinates as is when filling. Adding the delta when filling makes the fill off-by-one,
and this prevents us from adding the delta globally. The add only needs to be done when
antialiasing is off. Unfortunately the Quartz paint engine has trouble selecting the right
line width when drawing at non-integer coordinates, so we need to modify the line with
to make sure we get the correct behavior. Looks like this is causing problems when
drawing points, will investigate that later.

Cosmetic pens are simulated by setting the pen width to a value that corresponds to
the with of a pixel in user coordinates.

Task: related to 121704

[git-p4: depot-paths = "//depot/qt/main/": change = 232096]