p4i integration Integrate 234391 from 4.2 to main: Fix data stream binary compatibility issues with saveGeometry(), restoreGeometry()
Jasmin pointed out that it's a bad idea to stream bools and ints directly, so we now
use quint8 and qint32 instead. I've also added a magic number and set the 
QDataStream version. The version number is now in the major.minor format, so
that we can break backwards compatibility if we want to at some point.

Autotest: 234385

[git-p4: depot-paths = "//depot/qt/main/": change = 234392]